refactor: divide into namespaces
