+++ Server/Components/Pages/GamePage.razor
@@ -4,7 +4,7 @@
 @using Fracture.Server.Components.Layout
 @using Fracture.Server.Components.Popups
 @using Fracture.Server.Components.UI
-@using ItemGeneratorModels
+@using Fracture.Server.Modules.Items.Models
 
 @layout GameLayout
