+++ Server/Components/Popups/EquipmentPopup.razor
@@ -1,7 +1,8 @@
-﻿@using ExampleItemGenerator.Services.Generators
+﻿@using Fracture.Server.Components.UI
+@using Fracture.Server.Modules.Items.Models
+@using Fracture.Server.Modules.Items.Services
+
 @inject IItemGenerator ItemGenerator
-@using ItemGeneratorModels
-@using Fracture.Server.Components.UI
 
 <button class="btn btn-primary" @onclick="GenerateNewItem">Generate new item </button>
