+++ Server/Components/UI/ItemCard.razor
@@ -1,4 +1,4 @@
-@using ItemGeneratorModels
+@using Fracture.Server.Modules.Items.Models
 
 <div class="col">
     <div class="card" style="width: 15rem;">
