+++ Server/Components/UI/ItemStats.razor
@@ -1,4 +1,4 @@
-@using ItemGeneratorModels
+@using Fracture.Server.Modules.Items.Models
 
 <table class="table table-striped">
     @foreach (var itemStat in Enum.GetValues<ItemStat>())
