+++ Server/Components/UI/UserStats.razor
@@ -1,4 +1,4 @@
-﻿@using ItemGeneratorModels
+﻿@using Fracture.Server.Modules.Items.Models
 
 <table class="table">
     @foreach (var itemStat in Enum.GetValues<ItemStat>())
