+++ Server/Modules/Items/Services/ChristmasNameGenerator.cs
@@ -1,6 +1,7 @@
 ﻿using System;
+using Fracture.Server.Modules.Shared;
 
-namespace ExampleItemGenerator.Services.Generators
+namespace Fracture.Server.Modules.Items.Services
 {
     public class ChristmasNameGenerator : INameGenerator
     {
